Split out FieldRules into separate concrete classes
