+++ kubwa-compiler/src/main/java/com/wesleyelliott/kubwa/fieldrule/PasswordFieldRule.java
@@ -1,0 +1,12 @@
+package com.wesleyelliott.kubwa.fieldrule;
+
+import com.wesleyelliott.kubwa.rule.PasswordRule;
+
+/**
+ * Created by wesley on 2016/07/28.
+ */
+
+public class PasswordFieldRule extends FieldRule {
+
+    public PasswordRule.Scheme passwordScheme;
+}
