+++ kubwa-compiler/src/main/java/com/wesleyelliott/kubwa/fieldrule/RegexFieldRule.java
@@ -1,0 +1,10 @@
+package com.wesleyelliott.kubwa.fieldrule;
+
+/**
+ * Created by wesley on 2016/07/28.
+ */
+
+public class RegexFieldRule extends FieldRule {
+
+    public String regex;
+}
